rag-system/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Frontend
streamlit==1.31.0

# LLM & Embeddings
openai==1.12.0
tiktoken==0.5.2

# LangChain Orchestration
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Text Processing
sentence-transformers==2.3.1

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Testing (Development)
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Optional: Evaluation
# ragas==0.1.1  # Uncomment for automated evaluation
